Support execution of CREATE MATERIALIZED VIEW
Add createMaterializedView API to MetadataManager and add CreateMaterializedViewTask
that calls the API to perform MV creation.

HiveMetadata implementation of the API is to add essential MV parameters
and create MV as a standard table. ConnectorMaterializedViewDefinition is the json
structure that contains all essential metadata we will save to Metastore. It will
serialize into the viewOriginalText field of table metadata. Add MV partition.
At the moment, we only support partitioned MV defined on partitioned base tables.
MV must have one partition directly matched (selected) from base tables.

Add MATERIALIZED_VIEW table type.

We don't support alter schema of materialized view at the moment.

We don't support MVs defined across different catalogs, which is enforced in
CreateMaterializedViewTask

+++ presto-hive-metastore/src/main/java/com/facebook/presto/hive/metastore/SemiTransactionalHiveMetastore.java
@@ -78,6 +78,7 @@ import static com.facebook.presto.hive.metastore.MetastoreUtil.pathExists;
 import static com.facebook.presto.hive.metastore.MetastoreUtil.renameFile;
 import static com.facebook.presto.hive.metastore.MetastoreUtil.toPartitionValues;
 import static com.facebook.presto.hive.metastore.PrestoTableType.MANAGED_TABLE;
+import static com.facebook.presto.hive.metastore.PrestoTableType.MATERIALIZED_VIEW;
 import static com.facebook.presto.hive.metastore.PrestoTableType.TEMPORARY_TABLE;
 import static com.facebook.presto.hive.metastore.PrestoTableType.VIRTUAL_VIEW;
 import static com.facebook.presto.hive.metastore.Statistics.ReduceOperator.SUBTRACT;
@@ -501,7 +502,7 @@ public class SemiTransactionalHiveMetastore
         if (!table.isPresent()) {
             throw new TableNotFoundException(schemaTableName);
         }
-        if (!table.get().getTableType().equals(MANAGED_TABLE)) {
+        if (!table.get().getTableType().equals(MANAGED_TABLE) && !table.get().getTableType().equals(MATERIALIZED_VIEW)) {
             throw new PrestoException(NOT_SUPPORTED, "Cannot delete from non-managed Hive table");
         }
         if (!table.get().getPartitionColumns().isEmpty()) {
@@ -1148,7 +1149,7 @@ public class SemiTransactionalHiveMetastore
                 return;
             }
 
-            if (table.getTableType().equals(MANAGED_TABLE)) {
+            if (table.getTableType().equals(MANAGED_TABLE) || table.getTableType().equals(MATERIALIZED_VIEW)) {
                 String targetLocation = table.getStorage().getLocation();
                 checkArgument(!targetLocation.isEmpty(), "target location is empty");
                 Optional<Path> currentPath = tableAndMore.getCurrentLocation();
